requirements.txt
Django==5.1.1
djangorestframework==3.15.2
psycopg2==2.9.7
drf-spectacular==0.27.0
Pillow==10.0.1
uwsgi==2.0.22
flake8>=7.1.1
python-dotenv==1.0.0 
djangorestframework-simplejwt==5.3.0